fix(pipelines): drop float() casts in nested-SGLD grad fns
The float() casts on the metrics dict values broke under lax.scan tracing
in core.nested_sgld._warmup_step. Returning the traced jnp scalars works
in both the eager outer loop and the JIT-fused warmup chunk, since the
outer loop's float(m.get('hamiltonian')) call coerces a concrete array
back to a Python float without issue.

Verified end-to-end: model-form γ=1.0 smoke test now runs at ~700 steps/s
(vs 7 steps/s before) and returns sensible β posteriors near 1e4 instead
of collapsing to 1.0.

+++ src/pipelines/phase_b_model_form.py
@@ -126,7 +126,8 @@ def _build_grad_fns(
             return 0.5 * inv_var * jnp.sum((pred - y_obs) ** 2)
 
         nll_val, nll_grad = jax.value_and_grad(_nll)(phi_arr)
-        return float(nll_val), nll_grad
+        # Keep traced (no float cast) so this works inside lax.scan / jit.
+        return nll_val, nll_grad
 
     def prior_grad_fn(phi, lam, key):
         beta = jnp.exp(lam[0])
@@ -138,7 +139,7 @@ def _build_grad_fns(
         grad_phi = beta * g_phys
         # d/d(log beta) [beta * U] = beta * U
         grad_lam = jnp.array([beta * e_phys])
-        ham = float(beta * e_phys)
+        ham = beta * e_phys
 
         return grad_phi, grad_lam, {"hamiltonian": ham, "physics": e_phys}
 
@@ -153,7 +154,7 @@ def _build_grad_fns(
         grad_phi = beta * g_phys + g_like
         # Likelihood doesn't depend on beta, so grad_lam same as prior
         grad_lam = jnp.array([beta * e_phys])
-        ham = float(beta * e_phys + e_like)
+        ham = beta * e_phys + e_like
 
         return grad_phi, grad_lam, {
             "hamiltonian": ham,
